requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.6
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.1
executing==2.0.1
idna==3.7
iniconfig==2.0.0
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.1
wcwidth==0.2.13
